Noticket - Relax export plan serializer validation

+++ exportplan/core/serializers.py
@@ -64,10 +64,7 @@ class TargetMarketsResearchSerializer(serializers.Serializer):
     competitors = serializers.CharField(required=False, allow_blank=True, validators=[no_html])
     trend = serializers.CharField(required=False, allow_blank=True, validators=[no_html])
     unqiue_selling_proposition = serializers.CharField(required=False, allow_blank=True, validators=[no_html])
-    average_price = serializers.IntegerField(
-        required=False,
-        allow_null=True,
-    )
+    average_price = serializers.FloatField(required=False, allow_null=True)
 
 
 class RouteToMarketsSerializer(serializers.Serializer):
